move indexes to own directory in graph
