+++ wiggle_query_language/graph/indexes/tests/test_relationship_names.py
@@ -1,7 +1,7 @@
 from typing import Generator
 
 from testing import INDEXES_TEST_FILE_PATH
-from wiggle_query_language.graph.database.indexes.relationship_names_index import (
+from wiggle_query_language.graph.indexes.relationship_names_index import (
     add_items_to_relationship_names_index,
     load_relationship_names_index,
 )
